feat(models): define models

+++ internal/models/deposit.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Deposit struct {
-	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	ProductID uint      `gorm:"not null" json:"product_id"`
+	ID        int64      `gorm:"primaryKey;autoIncrement" json:"id"`
+	ProductID int64      `gorm:"not null" json:"product_id"`
 	Product   Product   `gorm:"foreignKey:ProductID;references:ID" json:"-"`
 	Stock     float64   `gorm:"not null;default:0" json:"stock"`
 	CreatedAt time.Time `gorm:"autoCreateTime:milli" json:"created_at"`
